Rewrite to use private queues in each direction
This involves a fair amount of faff, including:
 - having a separate channel for 'accepting' connections

One major problem is that handshakes can become stale if they are not
fully played out, which will stall the listener, since it'll wait for
an EOF that never comes. There needs to be some detection of whether
the connection is still waiting.