Move towards listen/connect symmetry
Use `readable` and `writable` in the same positions in listener and
client; both now echo an EOF if we get an EOF on the in queue, meaning
you can ^D from either to drop the connection (and the listener might
keep listening for another connection).

`-k` is slightly broken still, due to the latch not being refreshed
when a new connection comes along.